+++ example/lib/pages/chat_list_demo.dart
@@ -19,8 +19,12 @@ class ChatListDemo extends StatefulWidget {
 }
 
 class _ChatListDemoState extends State<ChatListDemo> {
-  final List<Color> colors = List<Color>();
-  final List<String> sessions = ["I'm session", "I'm session", "I'm session"];
+  final List<Color> colors = <Color>[];
+  final List<String> sessions = <String>[
+    'I\'m session',
+    'I\'m session',
+    'I\'m session'
+  ];
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -101,7 +105,7 @@ class _ChatListDemoState extends State<ChatListDemo> {
             sessions.insert(0, "I'm session");
           });
         },
-        child: Icon(Icons.add),
+        child: const Icon(Icons.add),
       ),
     );
   }