+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## [2.0.1]
 
-* fix viewport wrong index for SliverGrid #4.
+* fix closeToTrailing issue. extended_list#6
 
 ## [2.0.0]